Refactor table writers with generic Ipc<B> pattern and fix compilation
Introduce a shared `Ipc<B: Build>` writer that encapsulates the
flush → batch → write → reset lifecycle for all three table writers
(wavelengths, measurements, intensities). This replaces the previous
per-table StreamWriter boilerplate with a single generic implementation.

Key changes:
- Add `Build` trait and `Ipc<B>` struct in `writer.rs`
- Extract `util.rs` with shared Arrow decoding helpers (`col`, `batches`,
  `file_batches`, `nullable`)
- Remove `Buf`, `EOS`, `commit()`, `terminated()`, and `reopen()` — their
  behaviour is absorbed by `take_bytes()` (which includes stream finish)
  and a unified `open()` that restores writer state from existing segments
- Add `Dataset::new` dispatch: creates if path absent, opens if present
- Move `finished` flag from manifest TOML into the binary file header
- Simplify measurements record: `integration` is now `u32` (was `u64`)
- Make `Wavelengths`, `Measurements`, `Intensities` structs `pub` so
  `Dataset` accessors can return `&mut` references
- Fix Cargo.toml clippy lint group priorities

+++ src/measurements/record.rs
@@ -32,24 +32,13 @@ pub struct Record {
     /// C coordinate in SI base units, or `None` if unused.
     pub c: Option<f32>,
     /// Integration time in microseconds.
-    pub integration: u64,
+    pub integration: u32,
 }
 
 /* ----------------------------------------------------------------------- Trait Implementations */
 
 impl Display for Record {
-    /// Formats the value using the given formatter.
-    ///
-    /// # Display or Debug?
-    ///
-    /// Use [`Display`] to show measurement ID and timestamp; the minimum information required
-    /// to uniquely identify a record. Use [`Debug`] to show the full record, including coordinates
-    /// and integration time. See trait-level documentation for more information.
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Record {{ id: {}, timestamp: {} }}",
-            self.id, self.timestamp
-        )
+        write!(f, "Record {{ id: {}, timestamp: {} }}", self.id, self.timestamp)
     }
 }
